third_party/WebKit/LayoutTests/external/wpt/html/browsers/browsing-the-web/unloading-documents/navigation-within-beforeunload-expected.txt

This is a testharness.js-based test.
FAIL Triggering navigation from within beforeunload event assert_equals: expected 1000 but got 1
Harness: the test ran to completion.
